+++ clang-tools-extra/test/clang-tidy/checkers/bugprone/argument-comment-init-list-cxx20.cpp
@@ -5,10 +5,18 @@
 // RUN: %check_clang_tidy -check-suffix=TYPED -std=c++20-or-later %s bugprone-argument-comment %t -- \
 // RUN:   -config="{CheckOptions: { \
 // RUN:     bugprone-argument-comment.CommentTypedInitLists: true}}" --
+// RUN: %check_clang_tidy -check-suffix=TEMP -std=c++20-or-later %s bugprone-argument-comment %t -- \
+// RUN:   -config="{CheckOptions: { \
+// RUN:     bugprone-argument-comment.CommentParenthesizedTemporaries: true}}" --
 // RUN: %check_clang_tidy -check-suffix=BOTH -std=c++20-or-later %s bugprone-argument-comment %t -- \
 // RUN:   -config="{CheckOptions: { \
 // RUN:     bugprone-argument-comment.CommentAnonymousInitLists: true, \
 // RUN:     bugprone-argument-comment.CommentTypedInitLists: true}}" --
+// RUN: %check_clang_tidy -check-suffix=ALL -std=c++20-or-later %s bugprone-argument-comment %t -- \
+// RUN:   -config="{CheckOptions: { \
+// RUN:     bugprone-argument-comment.CommentAnonymousInitLists: true, \
+// RUN:     bugprone-argument-comment.CommentTypedInitLists: true, \
+// RUN:     bugprone-argument-comment.CommentParenthesizedTemporaries: true}}" --
 
 struct T {
   int value;
@@ -31,14 +39,21 @@ void test_designated_init() {
   // CHECK-FIXES-ANON: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
   // CHECK-MESSAGES-TYPED: warning: argument name 'dim' in comment does not match parameter name 'dims'
   // CHECK-FIXES-TYPED: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
+  // CHECK-MESSAGES-TEMP: warning: argument name 'dim' in comment does not match parameter name 'dims'
+  // CHECK-FIXES-TEMP: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
   // CHECK-MESSAGES-BOTH: warning: argument name 'dim' in comment does not match parameter name 'dims'
   // CHECK-FIXES-BOTH: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
+  // CHECK-MESSAGES-ALL: warning: argument name 'dim' in comment does not match parameter name 'dims'
+  // CHECK-FIXES-ALL: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
 
   foo_designated(some_arg, Agg{.x = 1});
-  // CHECK-MESSAGES-OFF-NOT: :[[@LINE-1]]:28: warning: argument comment missing for literal argument 'dims'
-  // CHECK-MESSAGES-ANON-NOT: :[[@LINE-2]]:28: warning: argument comment missing for literal argument 'dims'
-  // CHECK-MESSAGES-TYPED: [[@LINE-3]]:28: warning: argument comment missing for literal argument 'dims' [bugprone-argument-comment]
+  // CHECK-MESSAGES-OFF-NOT: :[[@LINE-1]]:28: warning: argument comment missing for argument 'dims'
+  // CHECK-MESSAGES-ANON-NOT: :[[@LINE-2]]:28: warning: argument comment missing for argument 'dims'
+  // CHECK-MESSAGES-TYPED: [[@LINE-3]]:28: warning: argument comment missing for argument 'dims' [bugprone-argument-comment]
   // CHECK-FIXES-TYPED: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
-  // CHECK-MESSAGES-BOTH: [[@LINE-5]]:28: warning: argument comment missing for literal argument 'dims' [bugprone-argument-comment]
+  // CHECK-MESSAGES-TEMP-NOT: :[[@LINE-5]]:28: warning: argument comment missing for argument 'dims'
+  // CHECK-MESSAGES-BOTH: [[@LINE-6]]:28: warning: argument comment missing for argument 'dims' [bugprone-argument-comment]
   // CHECK-FIXES-BOTH: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
+  // CHECK-MESSAGES-ALL: [[@LINE-8]]:28: warning: argument comment missing for argument 'dims' [bugprone-argument-comment]
+  // CHECK-FIXES-ALL: foo_designated(some_arg, /*dims=*/Agg{.x = 1});
 }